pyproject.toml
[project]
name = "archival-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "inngest>=0.5.16",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "numpy>=2.4.2",
    "orjson>=3.11.7",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
]
